Add Gemfile.lock and update .gitignore
I noticed that scheduled CI jobs weren't picking up up-to-date RuboCop
versions. The minimum gemfile jobs are working, but I'd expect the
standard run would grab the latest and greatest. Not so since, absent a
`Gemfile.lock`, the CI job is going to fall back to the cache. That has
the older gem versions.

Looking back at Shopify/ruby-style-guide [1] (which served as a
reference point for this testing strategy in the first place), I found
that they check in `Gemfile.lock`, have Dependabot do regular updates,
and their setup with current dependency versions and minimum versions
seems to be working just fine.

So, we're gonna give that a shot.

[1] https://github.com/Shopify/ruby-style-guide