Configure `Capybara/ClickLinkOrButtonStyle`
This rule's default enforce style, `link_or_button` [1], introduced a
number of problems [2]. Specifically, I noticed issues switching to
`click_on` from `click_button` [3]. The hack solution is to ignore or
disable the rule in some places but not in others. Not great.

As a result, the rule is slated for removal [4] and replacement [5] by
a new rule. In the interim, configuring this rule to enforce a strict
use of `click_button` and `click_link` is preferential (and, honestly,
more communicative).

In this author's opinion, user interface specs should be explicit about
the markup under test.

[1] https://docs.rubocop.org/rubocop-capybara/cops_capybara.html#capybaraclicklinkorbuttonstyle
[2] https://github.com/rubocop/rubocop-capybara/pull/61
[3] https://github.com/rubocop/rubocop-capybara/pull/61#issuecomment-1889733809
[4] https://github.com/rubocop/rubocop-capybara/issues/81
[5] https://github.com/rubocop/rubocop-capybara/pull/99