Test old versions with `Gemfile.local`
Issues in #82 reveal a weird challenge with the existing `gemfiles/` folder structure. Constants appear to not be loading for… reasons?

Some other related RuboCop gems take this tact of a `Gemfile.local` written by a CI task and conditionally eval'ed in the main Gemfile. Seems reasonable enough. Let's give it a try.